requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Async HTTP
aiohttp>=3.9.0
httpx>=0.26.0

# API Clients
google-api-python-client>=2.116.0
google-generativeai>=0.3.2
spotipy>=2.23.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
